+++ prisma/migrations/20260810080000_checkliste/migration.sql
@@ -1,0 +1,4 @@
+-- Zeitpunkte fuer die Veroeffentlichungs-Checkliste.
+ALTER TABLE "Giveaway" ADD COLUMN "termsPublishedAt" DATETIME;
+ALTER TABLE "Giveaway" ADD COLUMN "proofPublishedAt" DATETIME;
+ALTER TABLE "Giveaway" ADD COLUMN "lastUploadAt" DATETIME;